Fixed compiler warnings

+++ storage/connect/inihandl.c
@@ -152,7 +152,7 @@ static void PROFILE_CopyEntry( char *buffer, const char *value, uint len,
       if (!p2)
         continue;  /* ignore it */
 
-      strncpy(env_val, p + 2, min(sizeof(env_val), (int)(p2-p)-1));
+      strncpy(env_val, p + 2, min((int) sizeof(env_val), (int)(p2-p)-1));
 
       if ((env_p = getenv(env_val)) != NULL) {
         int buffer_len;
